+++ src/tcp/mod.rs
@@ -1,3 +1,4 @@
 pub mod active_session;
+pub mod frame_receiver;
 pub mod local_handler;
 pub mod state;
